requirements.txt
# Core dependencies
PySide6>=6.4.0
python-hcl2>=4.3.0

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-qt>=4.2.0
black>=22.0.0
mypy>=0.990
